一波更新,调整 Page 的两个方法名,以致可以 json 化,逻辑删除的与未删除的默认值交换一下

+++ mybatis-plus/src/test/java/com/baomidou/mybatisplus/test/base/entity/LogicTestData.java
@@ -1,12 +1,17 @@
 package com.baomidou.mybatisplus.test.base.entity;
 
-import com.baomidou.mybatisplus.annotation.*;
-import lombok.Data;
-import lombok.experimental.Accessors;
-
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+
+import com.baomidou.mybatisplus.annotation.FieldFill;
+import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableLogic;
+import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.annotation.Version;
+
+import lombok.Data;
+import lombok.experimental.Accessors;
 
 /**
  * @author meimie
@@ -34,5 +39,5 @@ public class LogicTestData {
     private Boolean deleted;
 
     @Version
-    private Long version;
+    private Integer version;
 }
